requirements.txt
fastapi
uvicorn
pandas
mysql-connector-python==8.0.33